backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
opencv-python==4.8.1.78
ultralytics==8.0.196
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
pydantic==2.5.0
python-dotenv==1.0.0
aiofiles==23.2.1

# OCR Libraries
easyocr==1.7.0
pytesseract==0.3.10

# Text-to-Speech
pyttsx3==2.90
gtts==2.4.0
playsound==1.3.0

# Text Processing
textblob==0.17.1

# GUI Framework (PyQt5)
PyQt5==5.15.10
